+++ core-tests/src/test/java/org/mule/runtime/core/processor/strategy/WorkQueueProcessingStrategyTestCase.java
@@ -34,7 +34,6 @@ import org.mule.runtime.core.transaction.TransactionCoordination;
 import org.mule.tck.testmodels.mule.TestTransaction;
 
 import org.junit.Test;
-
 import ru.yandex.qatools.allure.annotations.Description;
 import ru.yandex.qatools.allure.annotations.Features;
 import ru.yandex.qatools.allure.annotations.Stories;
