+++ tests/performance/src/test/java/org/mule/test/core/FlowProcessBenchmarkAssertionTestCase.java
@@ -7,13 +7,12 @@
 package org.mule.test.core;
 
 import static java.util.Collections.singletonMap;
-import static java.util.concurrent.TimeUnit.MICROSECONDS;
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
 
 import org.mule.AbstractBenchmarkAssertionTestCase;
 import org.mule.FlowNullProcessorBenchmark;
 import org.mule.runtime.core.processor.strategy.BlockingProcessingStrategyFactory;
-import org.mule.runtime.core.processor.strategy.DefaultFlowProcessingStrategyFactory;
+import org.mule.runtime.core.processor.strategy.TransactionAwareWorkQueueProcessingStrategyFactory;
 
 import org.junit.Test;
 
@@ -25,7 +24,7 @@ public class FlowProcessBenchmarkAssertionTestCase extends AbstractBenchmarkAsse
   public void processStreamOf1000FlowDefault() {
     runAndAssertBenchmark(FlowNullProcessorBenchmark.class, "processSourceStream", 1,
                           singletonMap(PROCESSING_STRATEGY_PARAM,
-                                       new String[] {DefaultFlowProcessingStrategyFactory.class.getCanonicalName()}),
+                                       new String[] {TransactionAwareWorkQueueProcessingStrategyFactory.class.getCanonicalName()}),
                           9, MILLISECONDS, 7800000);
   }
